Branching respects the sign of the relative address

+++ lib/instruction/bcc.c
@@ -1,7 +1,8 @@
+#include <stdint.h>
 #include "bcc.h"
 
 void bcc(struct cpu *cpu) {
 	if (!cpu->carry) {
-		cpu->program_counter += cpu->operand;
+		cpu->program_counter += (int8_t)cpu->operand;
 	}
 }
