+++ lib/instruction/beq.c
@@ -1,7 +1,8 @@
+#include <stdint.h>
 #include "beq.h"
 
 void beq(struct cpu *cpu) {
 	if (cpu->zero) {
-		cpu->program_counter += cpu->operand;
+		cpu->program_counter += (int8_t)cpu->operand;
 	}
 }
